aid/requirements.txt
numpy>=1.24.3
matplotlib>=3.8.0
seaborn>=0.12.2
evalplus>=0.2.0
pandas>=2.2.1
python-dotenv
# notebook